add a HTTP listener to receive data from ruuvi tags
This assumes that bluewalker [1] will be running and sending newline
separated JSON to a socket, and something else will be taking each line
and submitting to this new HTTP handler.

It's rube golberg-y, however bluewalker runs as root and I want a nice
big gap between it and this program.

[1] https://gitlab.com/jtaimisto/bluewalker