avoid Fatal errors in adapter code
if on adapter fails, we want that adapter to bail and log an error
without stopping the entire process.

Added a basic logging mechanism via pubsub, so eventually all logging
can be standardised and performed on a single goroutine. Maybe also with
log levels so I can skipping debugging output when it's not required.