untangle (mostly) the ruuvi adapter from http serving
Previously the ruuvi adapter was a special case, because it had to be
mounted as a http handler.

This decouples them. The http server now runs in a dedicated goroutine
and publishes the request details to the bus. It waits up to 2 seconds
for a response to appear on the bus, otherwise it returns an error to
the user.

This will do interesting things to memory usage an efficiency, but it's
a much neater solution. It should allow all the adapters to follow a
consistent pattern, and efficiency isn't my top concern.