Ensure unused subscriptions and topics are removed from the bus
This fixes an unbounded memory problem. Previously every new topic and
subscription was created in memory but never removed, even if the
receiver has gone away and stopped listening.

The result is more verbose, but obviously avoiding unbounded memory
growth is desirable.

The key change is the bus.Subscribe() now returns a struct with a
Close() method, which should be called via defer:

    subEveryMinute, _ := bus.Subscribe("every:minute")
    defer subEveryMinute.Close()
    for event := range subEveryMinute.Ch {
      // do things with event
    }