switch from in-memory state to cockroachdb
this is a very silly thing to do. Long term I don't want my monitoring
and automation to rely on an Internet connection. However I was
interested in learning about cockroachdb and this felt like an easy way
to do that.

I'll revert back to local storage at some point.