experiment with a shared package for updating state in a common way
Most of the state I store is monitoring a sensor. Mostly a numeric gauge
that varies over time, but sometimes a boolean on/off value. There's
probably others, but these are a good start.

Standardising the way things are stored in state is nice, also this
reduces how much code needs to directly interact with the pubsub bus.