Change /ruuvi HTTP endpoint to parse POSTs from a ruuvi gateway
I've just picked up a Ruuvi Gateway (https://ruuvi.com/gateway/). It
sits on my network, listens to BLE advertisements, and periodically POSTS
them to a HTTP server of my choice.

This converts the /ruuvi path to handle the shape of that request,
rather than the home brew thing I had that was using bluewalker to scan
for BLE advertisements, then writing the JSON output to a socket that
passed it to a program that POSTed to my server.

The gateway means I have an extra piece of hardware to look after, but
it takes care of all of the BLE mess for me a nd I can just deal with
easy HTTP POSTs.

I've included a sample JSON file with what the payload from the Ruuvi
Gateway looks like. Some of the advertisements are my ruuvitags, the
others are god knows what. Presumably phones and TVs and watches around
my house. For now I'm only processing the ruuvitags, but it'd be
interesting to consider what else I  could do with the other ones.

Sadly I had to vendor a package that helps with parsing the BLE data.

Long term, it'd be nice if this package was just responsible for parsing
the POST and data, then popping it onto the event bus. Other packages
could subscribe to the events and process ads for specific vendors, like
ruuvitags.