Switch to polling for ruuvitag data
I have a Ruuvi Gateway now, which monitors for BLE advertisments. We
have two options for getting the data:

1. the gateway can push it to us via HTTP requests
2. we can poll for it from a HTTP server on the gateway

Initially I was doing (1), because realtime. However, a recent
regression in the gateway firmware has stopped that work for me [1].
This switches to (2) - we poll the gateway for data.

The realtime nature of the push based approach is nice, but I don't
think it actually matters. When the bug in the firmware is fixed I might
revert this change, or I might just stick with it.

One slightly tricky thing with this was I had to disable auth on the
gateway to make polling work [2]. I think that's OK though, my network
is secure and even if it wasn't, someone monitoring the climate of the
house would be the least of my worries if they were on my network.

Finally, this means the HTTP server code has been ripped out of
home-data. I'm not opposed to running an HTTP server in this process,
I'm just removing it for now because it's not used any more.

[1] https://github.com/ruuvi/ruuvi.gateway_esp.c/issues/449
[2] https://f.ruuvi.com/t/how-can-i-enable-polling/5317